move LoggerTrait to Traits namespace
